refactor: handleDelete 함수를 MemberDeleteModal 내부로 이동
