+++ src/components/features/Project/MemberManagementModal/MemberDeleteModal.tsx
@@ -9,24 +9,63 @@ import {
   ModalHeader,
   ModalOverlay,
   Text,
+  useToast,
   VStack,
 } from "@chakra-ui/react";
+import type { AxiosError } from "axios";
+import { useParams } from "react-router-dom";
 
 import type { MemberResponseDTO } from "../../../../api/generated/data-contracts";
+import { useDeleteMember } from "../../../../api/hooks/useDeleteMember";
 import { MemberProfile } from "./MemberProfile";
 
-interface ProfileDeleteModalProps extends MemberResponseDTO {
+interface MemberDeleteModalProps extends MemberResponseDTO {
   isOpen: boolean;
   onClose: () => void;
-  onDelete: () => void;
 }
 
-export const ProfileDeleteModal = ({
+export const MemberDeleteModal = ({
   isOpen,
   onClose,
-  onDelete,
   ...member
-}: ProfileDeleteModalProps) => {
+}: MemberDeleteModalProps) => {
+  const projectId = useParams().id;
+  const mutation = useDeleteMember();
+  const toast = useToast();
+
+  const handleDelete = () => {
+    if (projectId && member.id) {
+      mutation.mutate(
+        { projectId: parseInt(projectId), memberId: member.id },
+        {
+          onSuccess: () => {
+            toast({
+              title: "팀원 삭제 성공",
+              description: `${member.name} 팀원이 삭제되었습니다.`,
+              status: "success",
+              duration: 3000,
+              isClosable: true,
+            });
+            onClose();
+          },
+          onError: (error: AxiosError) => {
+            const errorMessage = error.request.responseText;
+
+            toast({
+              title: "팀원 삭제 실패",
+              description: `${errorMessage}`,
+              status: "error",
+              duration: 3000,
+              isClosable: true,
+            });
+          },
+        }
+      );
+    } else {
+      console.error("프로젝트 ID 또는 멤버 ID가 없습니다.");
+    }
+  };
+
   return (
     <Modal isOpen={isOpen} onClose={onClose} size="sm" isCentered>
       <ModalOverlay />
@@ -44,7 +83,12 @@ export const ProfileDeleteModal = ({
 
         <ModalFooter>
           <Flex direction="column" flex={1} gap={3}>
-            <Button colorScheme="red" size="sm" onClick={onDelete} width="100%">
+            <Button
+              colorScheme="red"
+              size="sm"
+              onClick={handleDelete}
+              width="100%"
+            >
               삭제
             </Button>
             <Button variant="outline" size="sm" onClick={onClose} width="100%">
