+++ src/components/features/Project/MemberManagementModal/MemberItem.tsx
@@ -7,57 +7,19 @@ import {
   MenuItem,
   MenuList,
   useDisclosure,
-  useToast,
 } from "@chakra-ui/react";
-import type { AxiosError } from "axios";
 import { Ellipsis } from "lucide-react";
-import { useParams } from "react-router-dom";
 
 import type { MemberResponseDTO } from "../../../../api/generated/data-contracts";
-import { useDeleteMember } from "../../../../api/hooks/useDeleteMember";
-import { ProfileDeleteModal } from "./MemberDeleteModal";
+import { MemberDeleteModal } from "./MemberDeleteModal";
 import { MemberProfile } from "./MemberProfile";
 
-export const MemberItem = (member: MemberResponseDTO) => {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const mutation = useDeleteMember();
-  const projectId = useParams().id;
-  const toast = useToast();
-
-  const handleDelete = () => {
-    console.log(projectId);
-    console.log(member.id);
-    if (projectId && member.id) {
-      mutation.mutate(
-        { projectId: parseInt(projectId), memberId: member.id },
-        {
-          onSuccess: () => {
-            toast({
-              title: "팀원 삭제 성공",
-              description: `${member.name} 팀원이 삭제되었습니다.`,
-              status: "success",
-              duration: 3000,
-              isClosable: true,
-            });
-            onClose();
-          },
-          onError: (error: AxiosError) => {
-            const errorMessage = error.request.responseText;
-
-            toast({
-              title: "팀원 삭제 실패",
-              description: `${errorMessage}`,
-              status: "error",
-              duration: 3000,
-              isClosable: true,
-            });
-          },
-        }
-      );
-    } else {
-      console.error("프로젝트 ID 또는 멤버 ID가 없습니다.");
-    }
-  };
+export const MemberItem = ({ ...member }: MemberResponseDTO) => {
+  const {
+    isOpen: isMemberDeleteModalOpen,
+    onOpen: onMemberDeleteModalOpen,
+    onClose: onMemberDeleteModalClose,
+  } = useDisclosure();
 
   return (
     <Flex
@@ -85,7 +47,7 @@ export const MemberItem = (member: MemberResponseDTO) => {
           boxShadow="md"
           border="1px solid"
           borderColor="gray.100"
-          z-index={10}
+          zIndex={10}
         >
           <MenuItem textAlign="center" icon={<EditIcon />}>
             정보 수정
@@ -93,19 +55,19 @@ export const MemberItem = (member: MemberResponseDTO) => {
           <MenuItem
             textAlign="center"
             icon={<DeleteIcon />}
-            onClick={onOpen}
+            onClick={onMemberDeleteModalOpen}
             _hover={{ color: "red.500" }}
           >
             팀원 삭제
           </MenuItem>
-          <ProfileDeleteModal
-            isOpen={isOpen}
-            onClose={onClose}
-            onDelete={handleDelete}
-            {...member}
-          />
         </MenuList>
       </Menu>
+
+      <MemberDeleteModal
+        isOpen={isMemberDeleteModalOpen}
+        onClose={onMemberDeleteModalClose}
+        {...member}
+      />
     </Flex>
   );
 };
